+++ spec/helper/MockHelper.ts
@@ -267,7 +267,7 @@ interface EventEmitter {
 
   fireAdErrorEvent(): void;
 
-  fireVideoPlaybackQualityChangedEvent(bitrate: number): void;
+  fireVideoPlaybackQualityChangedEvent(bitrate: number, frameRate: number): void;
 
   fireCastStartedEvent(resuming?: boolean): void;
 
@@ -454,7 +454,7 @@ class PlayerEventHelper implements EventEmitter {
     });
   }
 
-  fireVideoPlaybackQualityChangedEvent(bitrate: number): void {
+  fireVideoPlaybackQualityChangedEvent(bitrate: number, frameRate: number): void {
     this.fireEvent<VideoPlaybackQualityChangedEvent>({
       timestamp: Date.now(),
       type: PlayerEvent.VideoPlaybackQualityChanged,
@@ -469,6 +469,7 @@ class PlayerEventHelper implements EventEmitter {
         bitrate: bitrate,
         width: null,
         height: null,
+        frameRate: frameRate,
       },
     });
   }